requirements.txt
requests>=2.31.0
google-genai>=0.3.0
tenacity>=8.2.0
